requirements.txt
pandas
numpy
matplotlib
scikit-learn 
pytest
pylint
pytest-cov
black
flask
openai==0.27.0
python-dotenv
nltk